chore: should not import UI in background
