+++ src/plugins/Gitcoin/define.tsx
@@ -2,7 +2,6 @@ import type { PluginConfig } from '../plugin'
 import { PreviewCard } from './PreviewCard'
 import React, { Suspense, useMemo, useState } from 'react'
 import { SnackbarContent } from '@material-ui/core'
-import { useExtensionPermission } from '../../components/DataSource/useExtensionPermission'
 import useSWR from 'swr'
 import { parseURL } from '../../utils/utils'
 import Services from '../../extension/service'
@@ -15,10 +14,10 @@ import { useValueRef } from '../../utils/hooks/useValueRef'
 import { useObservableValues } from '../../utils/hooks/useObservableMapSet'
 import type { GitcoinGrantMetadata } from './Services'
 import BigNumber from 'bignumber.js'
-import { EthereumTokenType, ERC20TokenRecord } from '../Wallet/database/types'
-import { getNetworkSettings } from '../Wallet/UI/Developer/SelectEthereumNetwork'
+import { EthereumTokenType } from '../Wallet/database/types'
 import { isNumber } from 'lodash-es'
 import { DonateSuccessDialog, DonateFailDialog } from './Dialogs'
+import { getNetworkSettings } from '../Wallet/UI/Developer/EthereumNetworkSettings'
 
 const isGitcoin = (x: string): boolean => x.startsWith('https://gitcoin.co/grants')
 export const GitcoinPluginDefine: PluginConfig = {
